Fixed the trapping calcualtion in OP.  It now passes the old test for trapping.  The new api accepts the old format.  Also moved some tests out of the test_script into more specify files in the integration folder.
