+++ test/integration/script_test.py
@@ -3,48 +3,6 @@ import scipy as sp
 import OpenPNM
 import pytest
 
-def test_cubic_standard_call():
-    pn = OpenPNM.Network.Cubic(shape=[3, 4, 5])
-    np.testing.assert_almost_equal(pn['pore.coords'][0], [0.5, 0.5, 0.5])
-
-
-def test_trim_extend():
-    pn = OpenPNM.Network.Cubic(shape=[5, 5, 5])
-    assert sp.all(sp.in1d(pn.find_neighbor_pores(pores=0), [1, 5, 25]))
-    assert [pn.Np, pn.Nt] == [125, 300]
-    pn.trim(pores=[0])
-    assert sp.all(sp.in1d(pn.find_neighbor_pores(pores=0), [1, 5, 25]))
-    assert [pn.Np, pn.Nt] == [124, 297]
-    pn.extend(pore_coords=[0, 0, 0], throat_conns=[[124, 0]])
-    assert [pn.Np, pn.Nt] == [125, 298]
-    assert sp.all(sp.in1d(pn.find_neighbor_pores(pores=0), [1, 5, 25, 124]))
-
-
-def test_stitch():
-    ctrl = OpenPNM.Base.Controller()
-    [Nx, Ny, Nz] = [10, 10, 10]
-    pn = OpenPNM.Network.Cubic(shape=[Nx, Ny, Nz])
-    pn2 = OpenPNM.Network.Cubic(shape=[Nx, Ny, Nz])
-    pn2['pore.coords'][:, 2] += Nz
-    pn.stitch(donor=pn2,
-              P_network=pn.pores('top'),
-              P_donor=pn2.pores('bottom'),
-              len_max=1,
-              method='nearest')
-    assert pn.Np == 2*pn2.Np  # Ensure number of pores doubled
-    assert pn.Nt == (2*pn2.Nt + Nx*Ny)  # Ensure correct number of new throats
-    assert pn2 not in ctrl.values()  # Donor Network is removed from Controller
-    # Reuse the donor Network in another stitch
-    pn2['pore.coords'][:, 2] -= 2*Nz
-    pn.stitch(donor=pn2,
-              P_network=pn.pores('bottom'),
-              P_donor=pn2.pores('top'),
-              len_max=1,
-              method='nearest')
-    assert pn.Np == 3*pn2.Np  # Ensure number of pores increased again
-    assert pn.Nt == (3*pn2.Nt + 2*Nx*Ny)  # Ensure correct number of new throats
-    ctrl.clear()
-
 
 def test_linear_solvers():
     pn = OpenPNM.Network.Cubic([1, 40, 30], spacing=0.0001)
@@ -225,42 +183,6 @@ def test_thermal_conduction():
     b = sp.reshape(b, (divs[0], divs[1]))
     diff = a - b
     assert sp.amax(np.absolute(diff)) < 0.015
-
-
-def test_OP_standard():
-    pn = OpenPNM.Network.Cubic(shape=[10, 10, 10], spacing=0.0001)
-    geom = OpenPNM.Geometry.Toray090(network=pn, pores=pn.Ps, throats=pn.Ts)
-    water = OpenPNM.Phases.Water(network=pn)
-    phys = OpenPNM.Physics.GenericPhysics(network=pn,
-                                          geometry=geom,
-                                          phase=water)
-    f = OpenPNM.Physics.models.capillary_pressure.washburn
-    phys.models.add(propname='throat.capillary_pressure',
-                    model=f)
-    OP = OpenPNM.Algorithms.OrdinaryPercolation(network=pn,
-                                                invading_phase=water)
-    OP.run(inlets=pn.pores('top'))
-    V_inv = sp.sum(pn['pore.volume'][OP['pore.inv_Pc'] < sp.inf])
-    V_tot = sp.sum(pn['pore.volume'])
-    assert V_inv/V_tot == 1.0
-
-
-def test_OP_trapping():
-    pn = OpenPNM.Network.Cubic(shape=[10, 10, 10], spacing=0.0001)
-    geom = OpenPNM.Geometry.Toray090(network=pn, pores=pn.Ps, throats=pn.Ts)
-    water = OpenPNM.Phases.Water(network=pn)
-    phys = OpenPNM.Physics.GenericPhysics(network=pn,
-                                          geometry=geom,
-                                          phase=water)
-    f = OpenPNM.Physics.models.capillary_pressure.washburn
-    phys.models.add(propname='throat.capillary_pressure',
-                    model=f)
-    OP = OpenPNM.Algorithms.OrdinaryPercolation(network=pn,
-                                                invading_phase=water)
-    OP.run(inlets=pn.pores('top'), outlets=pn.pores('bottom'), trapping=True)
-    V_inv = sp.sum(pn['pore.volume'][OP['pore.inv_Pc'] < sp.inf])
-    V_tot = sp.sum(pn['pore.volume'])
-    assert V_inv/V_tot < 0.95
 
 
 def test_Darcy_alg():
